rearranged/renamed create components, added inital desktop css
