+++ client/components/createMap.jsx
@@ -7,7 +7,7 @@ import '@mapbox/mapbox-gl-geocoder/dist/mapbox-gl-geocoder.css';
 mapboxgl.workerClass = MapboxWorker;
 mapboxgl.accessToken = process.env.MAPBOX_API_KEY;
 
-class Mapbox extends React.Component {
+export default class CreateMap extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -22,10 +22,57 @@ class Mapbox extends React.Component {
     };
     this.mapContainer = React.createRef();
     this.map = React.createRef();
-    this.handleClick = this.handleClick.bind(this);
-    this.getButtonText = this.getButtonText.bind(this);
-    this.handleChange = this.handleChange.bind(this);
-    this.handleLift = this.handleLift.bind(this);
+  }
+
+  handleClick = () => {
+    if (!this.state.location.name) {
+      return;
+    }
+    if (!this.state.inLocations) {
+      const newLocation = this.state.location;
+      this.setState({
+        locations: this.state.locations.concat(newLocation),
+        inLocations: true,
+        showFinish: false
+      }, () => setTimeout(() => {
+        this.setState({
+          showFinish: true
+        });
+      }, 2000));
+    } else {
+      window.alert(`${this.state.location.name.split(',')[0]} has already been added.`);
+    }
+  }
+
+  getButtonText = () => {
+    if (!this.state.location.name) {
+      return 'Search For a Location';
+    } else if (!this.state.inLocations || this.state.inLocations === null) {
+      return `Add ${this.state.location.name.split(',')[0]}`;
+    } else {
+      return `${this.state.location.name.split(',')[0]} Added!`;
+    }
+  }
+
+  handleChange = event => {
+    const name = event.target.value;
+    this.setState({
+      tripName: name
+    });
+  }
+
+  handleLift = () => {
+    event.preventDefault();
+    if (!this.state.locations[0]) {
+      document.activeElement.blur();
+      window.alert('Please add at least one location to your itinerary to continue.');
+    } else {
+      const trip = {
+        tripName: this.state.tripName,
+        locations: this.state.locations
+      };
+      this.props.onSubmit(trip);
+    }
   }
 
   componentDidMount() {
@@ -120,57 +167,6 @@ class Mapbox extends React.Component {
     this.map.current.remove();
   }
 
-  handleClick() {
-    if (!this.state.location.name) {
-      return;
-    }
-    if (!this.state.inLocations) {
-      const newLocation = this.state.location;
-      this.setState({
-        locations: this.state.locations.concat(newLocation),
-        inLocations: true,
-        showFinish: false
-      }, () => setTimeout(() => {
-        this.setState({
-          showFinish: true
-        });
-      }, 2000));
-    } else {
-      window.alert(`${this.state.location.name.split(',')[0]} has already been added.`);
-    }
-  }
-
-  getButtonText() {
-    if (!this.state.location.name) {
-      return 'Search For a Location';
-    } else if (!this.state.inLocations || this.state.inLocations === null) {
-      return `Add ${this.state.location.name.split(',')[0]}`;
-    } else {
-      return `${this.state.location.name.split(',')[0]} Added!`;
-    }
-  }
-
-  handleChange(event) {
-    const name = event.target.value;
-    this.setState({
-      tripName: name
-    });
-  }
-
-  handleLift() {
-    event.preventDefault();
-    if (!this.state.locations[0]) {
-      document.activeElement.blur();
-      window.alert('Please add at least one location to your itinerary to continue.');
-    } else {
-      const trip = {
-        tripName: this.state.tripName,
-        locations: this.state.locations
-      };
-      this.props.onSubmit(trip);
-    }
-  }
-
   render() {
     let addClass;
     let finishClass;
@@ -194,20 +190,13 @@ class Mapbox extends React.Component {
       : null;
     return (
       <>
+        {deletionNotice}
         <div className="map-box">
           <div ref={this.mapContainer} className="map-container" />
-          <button onClick={this.handleClick} className={addClass}>{this.getButtonText()}</button>
         </div>
-        <div className="map-form">
-          {deletionNotice}
-          <form onSubmit={this.handleLift} autoComplete="off">
-            <input className="name" required="required" type="text" name="trip-name" placeholder="Itinerary Name" onChange={this.handleChange} />
-            <input className={finishClass} type="submit" value="Finish Itinerary" />
-          </form>
-        </div>
+        <button onClick={this.handleClick} className={addClass}>{this.getButtonText()}</button>
+        <input className={finishClass} type="submit" value="Finish Itinerary" />
       </>
     );
   }
-}
-
-export default Mapbox;
+};
