+++ server/public/overides.css
@@ -45,6 +45,14 @@ a.mapboxgl-ctrl-logo {
   margin: 0 0 -4px 0;
 }
 
+.mapboxgl-ctrl-geocoder input {
+  font-size: 1rem;
+}
+
+.mapboxgl-ctrl-geocoder input:focus {
+  outline: none;
+}
+
 @media screen and (min-width: 415px) {
   .map-container {
     border-radius: 0.5rem;
@@ -83,5 +91,37 @@ a.mapboxgl-ctrl-logo {
 
   .mapboxgl-ctrl-top-right > .mapboxgl-ctrl {
     margin: 10px 0;
+  }
+}
+
+@media screen and (min-width: 900px) {
+  .mapboxgl-ctrl-top-right {
+    width: calc((100vw - 40px)/5 - 3px);
+    top: 175px;
+    left: calc(-20vw + -8px);
+    height: 41px;
+  }
+
+  .mapboxgl-ctrl-top-right > .mapboxgl-ctrl {
+    margin: 0;
+  }
+
+  .mapboxgl-ctrl-geocoder {
+    width: 100% !important;
+    border-radius: 0.5rem !important;
+    box-shadow: 0 1px 3px rgb(200, 200, 200) !important;
+  }
+
+  .mapboxgl-ctrl-geocoder input {
+    height: 100%;
+  }
+
+  .mapboxgl-ctrl-geocoder svg {
+    top: calc(50% - 10px);
+  }
+
+  .map-container {
+    overflow: visible;
+    position: unset;
   }
 }
