+++ server/public/styles.css
@@ -977,6 +977,52 @@ a:visited {
     padding: 0 20px;
   }
 
+  .create-page {
+    width: 100vw;
+    height: calc(100vh - 60px);
+    padding: 20px 0;
+    flex-direction: row;
+  }
+
+  .map-box {
+    height: 100%;
+    width: 75%;
+    margin-left: 10px;
+  }
+
+  .create-form-header {
+    color: #0e58a8;
+    border-bottom: solid thin #0e58a8;
+    padding-bottom: 5px;
+    margin-bottom: 20px;
+  }
+
+  .create-form-label,
+  .create-form-subheader {
+    color: #0e58a8;
+    font-size: 1.1rem;
+    font-weight: bold;
+  }
+
+  .create-form-subheader {
+    margin-top: 15px;
+  }
+
+  .create-form-name {
+    background-color: white;
+    border-radius: 0.5rem;
+    box-shadow: 0 1px 3px rgb(200, 200, 200);
+    font-size: 1rem;
+    color: #0e58a8;
+    padding: 10px 12.5px;
+    margin-top: 10px;
+    width: 100%;
+  }
+
+  .create-form-name::placeholder {
+    color: rgb(170, 170, 170);
+  }
+
   .location-page {
     display: flex;
     flex-direction: column;
@@ -1152,6 +1198,11 @@ a:visited {
     padding: 0 3px;
   }
 
+  .create-list {
+    padding: 0;
+    top: 125px;
+  }
+
   .trip-list-item {
     width: 100%;
     margin-bottom: 15px;
@@ -1162,7 +1213,8 @@ a:visited {
     width: calc(50% - 5px);
   }
 
-  .trip-list-module {
+  .trip-list-module,
+  .map-form {
     width: 25%;
     height: calc(100vh - 100px);
   }
@@ -1187,7 +1239,7 @@ a:visited {
 
   .itinerary-map-container {
     position: relative;
-    background-color: white;
+    background-color: #efe9e1;
     border-radius: 0.5rem;
     box-shadow: 0 1px 3px lightgray;
     width: 75%;
@@ -1208,9 +1260,11 @@ a:visited {
 
   .add,
   .finish {
-    max-width: 100%;
-    width: 30%;
-    left: 35%;
+    width: calc(20% - 2px);
+    height: 41px;
+    bottom: unset;
+    top: 255px;
+    left: 0;
     font-size: 1.2rem;
   }
 
@@ -1218,12 +1272,14 @@ a:visited {
     width: 50%;
   }
 
-  .trip-list-module {
+  .trip-list-module,
+  .map-form {
     width: 38%;
     margin-right: 10px;
   }
 
-  .itinerary-map-container {
+  .itinerary-map-container,
+  .map-box {
     width: 62%;
   }
 }
@@ -1235,11 +1291,13 @@ a:visited {
     max-width: 450px;
   }
 
-  .trip-list-module {
+  .trip-list-module,
+  .map-form {
     width: 30%;
   }
 
-  .itinerary-map-container {
+  .itinerary-map-container,
+  .map-box {
     width: 70%;
   }
 }
@@ -1251,11 +1309,13 @@ a:visited {
     width: calc(33% - 3.3px);
   }
 
-  .trip-list-module {
+  .trip-list-module,
+  .map-form {
     width: 25%;
   }
 
-  .itinerary-map-container {
+  .itinerary-map-container,
+  .map-box {
     width: 75%;
   }
 
@@ -1268,23 +1328,27 @@ a:visited {
 
 @media screen and (min-width: 1400px) {
 
-  .trip-list-module {
-      width: 20%;
-    }
+  .trip-list-module,
+  .map-form {
+    width: 20%;
+  }
 
-    .itinerary-map-container {
-      width: 80%;
-    }
+  .itinerary-map-container,
+  .map-box {
+    width: 80%;
+  }
 
 }
 
 @media screen and (min-width: 1800px) {
 
-  .trip-list-module {
+  .trip-list-module,
+  .map-form {
     width: 18%;
   }
 
-  .itinerary-map-container {
+  .itinerary-map-container,
+  .map-box {
     width: 82%;
   }
 
@@ -1292,11 +1356,13 @@ a:visited {
 
 @media screen and (min-width: 2000px) {
 
-  .trip-list-module {
+  .trip-list-module,
+  .map-form {
     width: 15%;
   }
 
-  .itinerary-map-container {
+  .itinerary-map-container,
+  .map-box {
     width: 85%;
   }
 
@@ -1304,11 +1370,13 @@ a:visited {
 
 @media screen and (min-width: 2560px) {
 
-  .trip-list-module {
+  .trip-list-module,
+  .map-form {
     max-width: 400px;
   }
 
-  .itinerary-map-container {
+  .itinerary-map-container,
+  .map-box {
     width: calc(100% - 390px);
   }
